* master:
  reshape with C ordering
  Add support for custom kkt solvers in cvxopt interface (#1064)
  SCIP interface (#1035)
  Big data bug fixes (#1074)
  Remove Gurobi conic interface bottleneck (#1065)
  flake8
  fixed #1066
  fixed #954
  outdated example
  Cache auxiliary output for some solvers (#1062)
  fixed complex diag bug

+++ cvxpy/reductions/solvers/conic_solvers/conic_solver.py
@@ -224,7 +224,7 @@ class ConicSolver(Solver):
             restructured_A.row = restructured_A.row.astype(np.int64)
             restructured_A.col = restructured_A.col.astype(np.int64)
             restructured_A = restructured_A.reshape(
-                restruct_mat.shape[0] * (problem.x.size + 1),
+                np.int64(restruct_mat.shape[0]) * (np.int64(problem.x.size) + 1),
                 problem.A.shape[1], order='F')
         else:
             restructured_A = problem.A
